Move Ctm reading files to the IO subfolder and namespace
